@@ -1,69 +1,53 @@
 {
-  "_from": "bson@^4.2.2",
-  "_id": "bson@4.6.1",
-  "_inBundle": false,
-  "_integrity": "sha512-I1LQ7Hz5zgwR4QquilLNZwbhPw0Apx7i7X9kGMBTsqPdml/03Q9NBtD9nt/19ahjlphktQImrnderxqpzeVDjw==",
-  "_location": "/bson",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "bson@^4.2.2",
-    "name": "bson",
-    "escapedName": "bson",
-    "rawSpec": "^4.2.2",
-    "saveSpec": null,
-    "fetchSpec": "^4.2.2"
-  },
-  "_requiredBy": [
-    "/mongodb",
-    "/mongoose"
+  "name": "bson",
+  "description": "A bson parser for node.js and the browser",
+  "keywords": [
+    "mongodb",
+    "bson",
+    "parser"
   ],
-  "_resolved": "https://registry.npmjs.org/bson/-/bson-4.6.1.tgz",
-  "_shasum": "2b5da517539bb0f7f3ffb54ac70a384ca899641c",
-  "_spec": "bson@^4.2.2",
-  "_where": "/home/lakshy/Desktop/Calling-app/backend/node_modules/mongoose",
+  "files": [
+    "lib",
+    "src",
+    "dist",
+    "bson.d.ts",
+    "etc/prepare.js",
+    "bower.json"
+  ],
+  "types": "bson.d.ts",
+  "version": "4.7.0",
   "author": {
     "name": "The MongoDB NodeJS Team",
     "email": "dbx-node@mongodb.com"
   },
-  "browser": {
-    "./lib/bson.js": "./dist/bson.browser.umd.js",
-    "./dist/bson.esm.js": "./dist/bson.browser.esm.js"
-  },
+  "license": "Apache-2.0",
+  "contributors": [],
+  "repository": "mongodb/js-bson",
   "bugs": {
     "url": "https://jira.mongodb.org/projects/NODE/issues/"
   },
-  "bundleDependencies": false,
-  "config": {
-    "native": false
-  },
-  "contributors": [],
-  "dependencies": {
-    "buffer": "^5.6.0"
-  },
-  "deprecated": false,
-  "description": "A bson parser for node.js and the browser",
   "devDependencies": {
     "@babel/plugin-external-helpers": "^7.10.4",
     "@babel/preset-env": "^7.11.0",
     "@istanbuljs/nyc-config-typescript": "^1.0.1",
-    "@microsoft/api-extractor": "^7.11.2",
+    "@microsoft/api-extractor": "^7.28.0",
     "@rollup/plugin-babel": "^5.2.0",
     "@rollup/plugin-commonjs": "^15.0.0",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^9.0.0",
+    "@rollup/plugin-replace": "^4.0.0",
     "@rollup/plugin-typescript": "^6.0.0",
-    "@typescript-eslint/eslint-plugin": "^3.10.1",
-    "@typescript-eslint/parser": "^3.10.1",
+    "@types/node": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.0",
+    "@typescript-eslint/parser": "^5.30.0",
     "array-includes": "^3.1.3",
+    "array.prototype.flatmap": "^1.3.0",
     "benchmark": "^2.1.4",
     "chai": "^4.2.0",
-    "downlevel-dts": "^0.7.0",
-    "eslint": "^7.7.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-tsdoc": "^0.2.6",
+    "eslint": "^8.18.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.1.0",
+    "eslint-plugin-tsdoc": "^0.2.16",
     "karma": "^6.3.4",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^3.1.0",
@@ -73,63 +57,20 @@
     "mocha": "5.2.0",
     "node-fetch": "^2.6.1",
     "nyc": "^15.1.0",
-    "object.entries": "^1.1.4",
-    "prettier": "^2.1.1",
+    "object.entries": "^1.1.5",
+    "prettier": "^2.7.1",
     "rimraf": "^3.0.2",
     "rollup": "^2.26.5",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-node-globals": "^1.4.0",
     "rollup-plugin-node-polyfills": "^0.2.1",
     "rollup-plugin-polyfill-node": "^0.7.0",
-    "standard-version": "^9.3.0",
+    "standard-version": "^9.5.0",
     "ts-node": "^9.0.0",
-    "tsd": "^0.17.0",
-    "typedoc": "^0.21.2",
-    "typescript": "^4.0.2",
+    "tsd": "^0.21.0",
+    "typescript": "^4.7.4",
     "typescript-cached-transpile": "0.0.6",
     "uuid": "^8.3.2"
-  },
-  "engines": {
-    "node": ">=6.9.0"
-  },
-  "files": [
-    "lib",
-    "src",
-    "dist",
-    "bson.d.ts",
-    "etc/prepare.js",
-    "bower.json"
-  ],
-  "homepage": "https://github.com/mongodb/js-bson#readme",
-  "keywords": [
-    "mongodb",
-    "bson",
-    "parser"
-  ],
-  "license": "Apache-2.0",
-  "main": "lib/bson.js",
-  "module": "dist/bson.esm.js",
-  "name": "bson",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mongodb/js-bson.git"
-  },
-  "scripts": {
-    "build": "npm run build:dts && npm run build:bundle",
-    "build:bundle": "rollup -c rollup.config.js",
-    "build:dts": "npm run build:ts && api-extractor run --typescript-compiler-folder node_modules/typescript --local && rimraf 'lib/**/*.d.ts*' && downlevel-dts bson.d.ts bson.d.ts",
-    "build:ts": "tsc",
-    "coverage": "nyc npm run test-node",
-    "coverage:html": "npm run coverage && open ./coverage/index.html",
-    "docs": "typedoc",
-    "format": "eslint --ext '.js,.ts' src test --fix",
-    "lint": "eslint -v && eslint --ext '.js,.ts' --max-warnings=0 src test && tsc -v && tsc --noEmit && npm run test-tsd",
-    "prepare": "node etc/prepare.js",
-    "release": "standard-version -i HISTORY.md",
-    "test": "npm run build && npm run test-node && npm run test-browser",
-    "test-browser": "node --max-old-space-size=4096 ./node_modules/.bin/karma start karma.conf.js",
-    "test-node": "mocha test/node test/*_tests.js",
-    "test-tsd": "npm run build:dts && tsd"
   },
   "tsd": {
     "directory": "test/types",
