@@ -1,93 +1,50 @@
 {
-  "_from": "http-errors@1.8.1",
-  "_id": "http-errors@1.8.1",
-  "_inBundle": false,
-  "_integrity": "sha512-Kpk9Sm7NmI+RHhnj6OIWDI1d6fIoFAtFt9RLaTMRlg/8w49juAStsrBgp0Dp4OdxdVbRIeKhtCUvoi/RuAhO4g==",
-  "_location": "/http-errors",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "http-errors@1.8.1",
-    "name": "http-errors",
-    "escapedName": "http-errors",
-    "rawSpec": "1.8.1",
-    "saveSpec": null,
-    "fetchSpec": "1.8.1"
-  },
-  "_requiredBy": [
-    "/body-parser",
-    "/raw-body",
-    "/send"
-  ],
-  "_resolved": "https://registry.npmjs.org/http-errors/-/http-errors-1.8.1.tgz",
-  "_shasum": "7c3f28577cbc8a207388455dbd62295ed07bd68c",
-  "_spec": "http-errors@1.8.1",
-  "_where": "/home/lakshy/Desktop/Calling-app/backend/node_modules/body-parser",
-  "author": {
-    "name": "Jonathan Ong",
-    "email": "me@jongleberry.com",
-    "url": "http://jongleberry.com"
-  },
-  "bugs": {
-    "url": "https://github.com/jshttp/http-errors/issues"
-  },
-  "bundleDependencies": false,
+  "name": "http-errors",
+  "description": "Create HTTP error objects",
+  "version": "2.0.0",
+  "author": "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)",
   "contributors": [
-    {
-      "name": "Alan Plum",
-      "email": "me@pluma.io"
-    },
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    }
+    "Alan Plum <me@pluma.io>",
+    "Douglas Christopher Wilson <doug@somethingdoug.com>"
   ],
+  "license": "MIT",
+  "repository": "jshttp/http-errors",
   "dependencies": {
-    "depd": "~1.1.2",
+    "depd": "2.0.0",
     "inherits": "2.0.4",
     "setprototypeof": "1.2.0",
-    "statuses": ">= 1.5.0 < 2",
+    "statuses": "2.0.1",
     "toidentifier": "1.0.1"
   },
-  "deprecated": false,
-  "description": "Create HTTP error objects",
   "devDependencies": {
     "eslint": "7.32.0",
     "eslint-config-standard": "14.1.1",
     "eslint-plugin-import": "2.25.3",
     "eslint-plugin-markdown": "2.2.1",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "5.1.1",
+    "eslint-plugin-promise": "5.2.0",
     "eslint-plugin-standard": "4.1.0",
     "mocha": "9.1.3",
     "nyc": "15.1.0"
   },
   "engines": {
-    "node": ">= 0.6"
-  },
-  "files": [
-    "index.js",
-    "HISTORY.md",
-    "LICENSE",
-    "README.md"
-  ],
-  "homepage": "https://github.com/jshttp/http-errors#readme",
-  "keywords": [
-    "http",
-    "error"
-  ],
-  "license": "MIT",
-  "name": "http-errors",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/http-errors.git"
+    "node": ">= 0.8"
   },
   "scripts": {
     "lint": "eslint . && node ./scripts/lint-readme-list.js",
     "test": "mocha --reporter spec --bail",
     "test-ci": "nyc --reporter=lcov --reporter=text npm test",
-    "test-cov": "nyc --reporter=html --reporter=text npm test"
+    "test-cov": "nyc --reporter=html --reporter=text npm test",
+    "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "1.8.1"
+  "keywords": [
+    "http",
+    "error"
+  ],
+  "files": [
+    "index.js",
+    "HISTORY.md",
+    "LICENSE",
+    "README.md"
+  ]
 }
