@@ -1,37 +1,40 @@
 {
-  "_from": "mongodb-connection-string-url@^2.4.1",
-  "_id": "mongodb-connection-string-url@2.5.2",
-  "_inBundle": false,
-  "_integrity": "sha512-tWDyIG8cQlI5k3skB6ywaEA5F9f5OntrKKsT/Lteub2zgwSUlhqEN2inGgBTm8bpYJf8QYBdA/5naz65XDpczA==",
-  "_location": "/mongodb-connection-string-url",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "mongodb-connection-string-url@^2.4.1",
-    "name": "mongodb-connection-string-url",
-    "escapedName": "mongodb-connection-string-url",
-    "rawSpec": "^2.4.1",
-    "saveSpec": null,
-    "fetchSpec": "^2.4.1"
-  },
-  "_requiredBy": [
-    "/mongodb"
+  "name": "mongodb-connection-string-url",
+  "version": "2.5.3",
+  "description": "MongoDB connection strings, based on the WhatWG URL API",
+  "keywords": [
+    "password",
+    "prompt",
+    "tty"
   ],
-  "_resolved": "https://registry.npmjs.org/mongodb-connection-string-url/-/mongodb-connection-string-url-2.5.2.tgz",
-  "_shasum": "f075c8d529e8d3916386018b8a396aed4f16e5ed",
-  "_spec": "mongodb-connection-string-url@^2.4.1",
-  "_where": "/home/lakshy/Desktop/Calling-app/backend/node_modules/mongodb",
+  "homepage": "https://github.com/mongodb-js/mongodb-connection-string-url",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/mongodb-js/mongodb-connection-string-url.git"
+  },
   "bugs": {
     "url": "https://github.com/mongodb-js/mongodb-connection-string-url/issues"
   },
-  "bundleDependencies": false,
-  "dependencies": {
-    "@types/whatwg-url": "^8.2.1",
-    "whatwg-url": "^11.0.0"
+  "main": "lib/index.js",
+  "exports": {
+    "require": "./lib/index.js",
+    "import": "./.esm-wrapper.mjs"
   },
-  "deprecated": false,
-  "description": "MongoDB connection strings, based on the WhatWG URL API",
+  "files": [
+    "LICENSE",
+    "lib",
+    "package.json",
+    "README.md",
+    ".esm-wrapper.mjs"
+  ],
+  "scripts": {
+    "lint": "eslint \"{src,test}/**/*.ts\"",
+    "test": "npm run lint && npm run build && nyc mocha --colors -r ts-node/register test/*.ts",
+    "build": "npm run compile-ts && gen-esm-wrapper . ./.esm-wrapper.mjs",
+    "prepack": "npm run build",
+    "compile-ts": "tsc -p tsconfig.json"
+  },
+  "license": "Apache-2.0",
   "devDependencies": {
     "@types/chai": "^4.2.5",
     "@types/mocha": "^8.0.3",
@@ -49,39 +52,11 @@
     "gen-esm-wrapper": "^1.1.3",
     "mocha": "^8.1.3",
     "nyc": "^15.1.0",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.3"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
   },
-  "exports": {
-    "require": "./lib/index.js",
-    "import": "./.esm-wrapper.mjs"
-  },
-  "files": [
-    "LICENSE",
-    "lib",
-    "package.json",
-    "README.md",
-    ".esm-wrapper.mjs"
-  ],
-  "homepage": "https://github.com/mongodb-js/mongodb-connection-string-url",
-  "keywords": [
-    "password",
-    "prompt",
-    "tty"
-  ],
-  "license": "Apache-2.0",
-  "main": "lib/index.js",
-  "name": "mongodb-connection-string-url",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mongodb-js/mongodb-connection-string-url.git"
-  },
-  "scripts": {
-    "build": "npm run compile-ts && gen-esm-wrapper . ./.esm-wrapper.mjs",
-    "compile-ts": "tsc -p tsconfig.json",
-    "lint": "eslint \"{src,test}/**/*.ts\"",
-    "prepack": "npm run build",
-    "test": "npm run lint && npm run build && nyc mocha --colors -r ts-node/register test/*.ts"
-  },
-  "version": "2.5.2"
+  "dependencies": {
+    "@types/whatwg-url": "^8.2.1",
+    "whatwg-url": "^11.0.0"
+  }
 }
