@@ -1,143 +1,127 @@
 {
-  "_from": "mongodb@4.3.1",
-  "_id": "mongodb@4.3.1",
-  "_inBundle": false,
-  "_integrity": "sha512-sNa8APSIk+r4x31ZwctKjuPSaeKuvUeNb/fu/3B6dRM02HpEgig7hTHM8A/PJQTlxuC/KFWlDlQjhsk/S43tBg==",
-  "_location": "/mongodb",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "mongodb@4.3.1",
-    "name": "mongodb",
-    "escapedName": "mongodb",
-    "rawSpec": "4.3.1",
-    "saveSpec": null,
-    "fetchSpec": "4.3.1"
-  },
-  "_requiredBy": [
-    "/mongoose"
-  ],
-  "_resolved": "https://registry.npmjs.org/mongodb/-/mongodb-4.3.1.tgz",
-  "_shasum": "e346f76e421ec6f47ddea5c8f5140e6181aaeb94",
-  "_spec": "mongodb@4.3.1",
-  "_where": "/home/lakshy/Desktop/Calling-app/backend/node_modules/mongoose",
-  "author": {
-    "name": "The MongoDB NodeJS Team",
-    "email": "dbx-node@mongodb.com"
-  },
-  "bugs": {
-    "url": "https://jira.mongodb.org/projects/NODE/issues/"
-  },
-  "bundleDependencies": false,
-  "dependencies": {
-    "bson": "^4.6.1",
-    "denque": "^2.0.1",
-    "mongodb-connection-string-url": "^2.4.1",
-    "saslprep": "^1.0.3",
-    "socks": "^2.6.1"
-  },
-  "deprecated": false,
+  "name": "mongodb",
+  "version": "4.8.1",
   "description": "The official MongoDB driver for Node.js",
-  "devDependencies": {
-    "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.19.4",
-    "@microsoft/tsdoc-config": "^0.15.2",
-    "@types/chai": "^4.3.0",
-    "@types/chai-subset": "^1.3.3",
-    "@types/kerberos": "^1.1.1",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^17.0.10",
-    "@types/saslprep": "^1.0.1",
-    "@types/semver": "^7.3.9",
-    "@types/sinon": "^10.0.6",
-    "@types/whatwg-url": "^8.2.1",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
-    "bluebird": "^3.7.2",
-    "chai": "^4.3.4",
-    "chai-subset": "^1.6.0",
-    "chalk": "^4.1.2",
-    "downlevel-dts": "^0.8.0",
-    "eslint": "^8.7.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
-    "eslint-plugin-tsdoc": "^0.2.14",
-    "js-yaml": "^4.1.0",
-    "lodash.camelcase": "^4.3.0",
-    "mocha": "^9.1.4",
-    "mocha-sinon": "^2.1.2",
-    "nyc": "^15.1.0",
-    "prettier": "^2.5.1",
-    "rimraf": "^3.0.2",
-    "semver": "^7.3.5",
-    "sinon": "^12.0.1",
-    "sinon-chai": "^3.7.0",
-    "source-map-support": "^0.5.21",
-    "standard-version": "^9.3.2",
-    "ts-node": "^10.4.0",
-    "tsd": "^0.19.1",
-    "typescript": "^4.5.4",
-    "typescript-cached-transpile": "^0.0.6",
-    "xml2js": "^0.4.23",
-    "yargs": "^17.3.1"
-  },
-  "engines": {
-    "node": ">=12.9.0"
-  },
+  "main": "lib/index.js",
   "files": [
     "lib",
     "src",
     "etc/prepare.js",
     "mongodb.d.ts",
-    "mongodb.ts34.d.ts"
+    "tsconfig.json"
   ],
-  "homepage": "https://github.com/mongodb/node-mongodb-native",
+  "types": "mongodb.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:mongodb/node-mongodb-native.git"
+  },
   "keywords": [
     "mongodb",
     "driver",
     "official"
   ],
+  "author": {
+    "name": "The MongoDB NodeJS Team",
+    "email": "dbx-node@mongodb.com"
+  },
+  "dependencies": {
+    "bson": "^4.6.5",
+    "denque": "^2.0.1",
+    "mongodb-connection-string-url": "^2.5.2",
+    "socks": "^2.6.2"
+  },
+  "devDependencies": {
+    "@iarna/toml": "^2.2.5",
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
+    "@microsoft/api-extractor": "^7.25.0",
+    "@microsoft/tsdoc-config": "^0.16.1",
+    "@mongodb-js/zstd": "^1.0.0",
+    "@types/chai": "^4.3.1",
+    "@types/chai-subset": "^1.3.3",
+    "@types/express": "^4.17.13",
+    "@types/kerberos": "^1.1.1",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^17.0.42",
+    "@types/saslprep": "^1.0.1",
+    "@types/semver": "^7.3.9",
+    "@types/sinon": "^10.0.11",
+    "@types/sinon-chai": "^3.2.8",
+    "@types/whatwg-url": "^8.2.1",
+    "@typescript-eslint/eslint-plugin": "^5.28.0",
+    "@typescript-eslint/parser": "^5.28.0",
+    "bluebird": "^3.7.2",
+    "chai": "^4.3.6",
+    "chai-subset": "^1.6.0",
+    "chalk": "^4.1.2",
+    "eslint": "^8.17.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "eslint-plugin-tsdoc": "^0.2.16",
+    "express": "^4.18.1",
+    "js-yaml": "^4.1.0",
+    "mocha": "^9.2.2",
+    "mocha-sinon": "^2.1.2",
+    "nyc": "^15.1.0",
+    "prettier": "^2.7.0",
+    "rimraf": "^3.0.2",
+    "semver": "^7.3.7",
+    "sinon": "^13.0.1",
+    "sinon-chai": "^3.7.0",
+    "source-map-support": "^0.5.21",
+    "standard-version": "^9.5.0",
+    "ts-node": "^10.8.1",
+    "tsd": "^0.21.0",
+    "typescript": "^4.7.3",
+    "typescript-cached-transpile": "^0.0.6",
+    "xml2js": "^0.4.23",
+    "yargs": "^17.5.1"
+  },
   "license": "Apache-2.0",
-  "main": "lib/index.js",
-  "name": "mongodb",
+  "engines": {
+    "node": ">=12.9.0"
+  },
+  "bugs": {
+    "url": "https://jira.mongodb.org/projects/NODE/issues/"
+  },
+  "homepage": "https://github.com/mongodb/node-mongodb-native",
   "optionalDependencies": {
     "saslprep": "^1.0.3"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/mongodb/node-mongodb-native.git"
-  },
   "scripts": {
-    "build:docs": "typedoc",
-    "build:dts": "npm run build:ts && api-extractor run && rimraf 'lib/**/*.d.ts*' && downlevel-dts mongodb.d.ts mongodb.ts34.d.ts",
     "build:evergreen": "node .evergreen/generate_evergreen_tasks.js",
-    "build:ts": "rimraf lib && ./node_modules/typescript/bin/tsc",
-    "check:adl": "mocha --file test/tools/runner test/manual/data_lake.test.js",
-    "check:atlas": "mocha --config \"test/manual/mocharc.json\" test/manual/atlas_connectivity.test.js",
-    "check:aws": "mocha --file test/tools/runner test/integration/auth/mongodb_aws.test.js",
+    "build:ts": "node ./node_modules/typescript/bin/tsc",
+    "build:dts": "npm run build:ts && api-extractor run && rimraf 'lib/**/*.d.ts*'",
+    "build:docs": "./etc/docs/build.ts",
+    "build:typedoc": "typedoc",
     "check:bench": "node test/benchmarks/driverBench",
     "check:coverage": "nyc npm run test:all",
-    "check:csfle": "mocha --file test/tools/runner test/integration/client-side-encryption",
-    "check:dts": "./node_modules/typescript/bin/tsc --noEmit mongodb.d.ts && tsd",
-    "check:eslint": "eslint -v && eslint --max-warnings=0 --ext '.js,.ts' src test",
-    "check:kerberos": "mocha --config \"test/manual/mocharc.json\" test/manual/kerberos.test.js",
-    "check:ldap": "mocha --config \"test/manual/mocharc.json\" test/manual/ldap.test.js",
+    "check:integration-coverage": "nyc npm run check:test",
     "check:lint": "npm run build:dts && npm run check:dts && npm run check:eslint && npm run check:tsd",
-    "check:ocsp": "mocha --config \"test/manual/mocharc.json\" test/manual/ocsp_support.test.js",
-    "check:snappy": "mocha test/unit/assorted/snappy.test.js",
-    "check:socks5": "mocha --config \"test/manual/mocharc.json\" test/manual/socks5.test.ts",
-    "check:test": "mocha --file test/tools/runner test/integration",
-    "check:tls": "mocha --config \"test/manual/mocharc.json\" test/manual/tls_support.test.js",
-    "check:ts": "./node_modules/typescript/bin/tsc -v && ./node_modules/typescript/bin/tsc --noEmit",
+    "check:eslint": "eslint -v && eslint --max-warnings=0 --ext '.js,.ts' src test",
     "check:tsd": "tsd --version && tsd",
-    "check:unit": "mocha test/unit/",
+    "check:dts": "node ./node_modules/typescript/bin/tsc --noEmit mongodb.d.ts && tsd",
+    "check:test": "mocha --config test/mocha_mongodb.json test/integration",
+    "check:unit": "mocha test/unit",
+    "check:ts": "node ./node_modules/typescript/bin/tsc -v && node ./node_modules/typescript/bin/tsc --noEmit",
+    "check:atlas": "mocha --config test/manual/mocharc.json test/manual/atlas_connectivity.test.js",
+    "check:adl": "mocha --config test/mocha_mongodb.json test/manual/atlas-data-lake-testing",
+    "check:aws": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_aws.test.ts",
+    "check:ocsp": "mocha --config test/manual/mocharc.json test/manual/ocsp_support.test.js",
+    "check:kerberos": "mocha --config test/manual/mocharc.json test/manual/kerberos.test.js",
+    "check:tls": "mocha --config test/manual/mocharc.json test/manual/tls_support.test.js",
+    "check:ldap": "mocha --config test/manual/mocharc.json test/manual/ldap.test.js",
+    "check:socks5": "mocha --config test/manual/mocharc.json test/manual/socks5.test.ts",
+    "check:csfle": "mocha --config test/mocha_mongodb.json test/integration/client-side-encryption",
+    "check:snappy": "mocha test/unit/assorted/snappy.test.js",
+    "fix:eslint": "npm run check:eslint -- --fix",
     "prepare": "node etc/prepare.js",
-    "release": "standard-version -i HISTORY.md",
+    "preview:docs": "ts-node etc/docs/preview.ts",
+    "release": "standard-version -a -i HISTORY.md",
     "test": "npm run check:lint && npm run test:all",
-    "test:all": "npm run check:unit && npm run check:test"
+    "test:all": "npm run check:unit && npm run check:test",
+    "update:docs": "npm run build:docs -- --yes"
   },
   "tsd": {
     "directory": "test/types",
