@@ -105,7 +18,103 @@
     "db"
   ],
   "license": "MIT",
+  "dependencies": {
+    "bson": "^4.6.5",
+    "kareem": "2.4.1",
+    "mongodb": "4.8.1",
+    "mpath": "0.9.0",
+    "mquery": "4.0.3",
+    "ms": "2.1.3",
+    "sift": "16.0.0"
+  },
+  "devDependencies": {
+    "@babel/core": "7.18.10",
+    "@typescript-eslint/eslint-plugin": "5.32.0",
+    "@typescript-eslint/parser": "5.32.0",
+    "acquit": "1.2.1",
+    "acquit-ignore": "0.2.0",
+    "acquit-require": "0.1.1",
+    "assert-browserify": "2.0.0",
+    "axios": "0.27.2",
+    "babel-loader": "8.2.5",
+    "benchmark": "2.1.4",
+    "bluebird": "3.7.2",
+    "buffer": "^5.6.0",
+    "cheerio": "1.0.0-rc.12",
+    "crypto-browserify": "3.12.0",
+    "dox": "0.9.1",
+    "eslint": "8.21.0",
+    "eslint-plugin-mocha-no-only": "1.1.1",
+    "express": "^4.18.1",
+    "highlight.js": "11.6.0",
+    "lodash.isequal": "4.5.0",
+    "lodash.isequalwith": "4.4.0",
+    "marked": "4.0.18",
+    "mkdirp": "^1.0.4",
+    "mocha": "10.0.0",
+    "moment": "2.x",
+    "mongodb-memory-server": "8.8.0",
+    "ncp": "^2.0.0",
+    "nyc": "15.1.0",
+    "pug": "3.0.2",
+    "q": "1.5.1",
+    "sinon": "14.0.0",
+    "stream-browserify": "3.0.0",
+    "ts-benchmark": "^1.1.10",
+    "tsd": "0.20.0",
+    "typescript": "4.7.4",
+    "uuid": "8.3.2",
+    "webpack": "5.74.0"
+  },
+  "directories": {
+    "lib": "./lib/mongoose"
+  },
+  "scripts": {
+    "docs:clean": "npm run docs:clean:stable",
+    "docs:clean:stable": "rimraf index.html && rimraf -rf ./docs/*.html  && rimraf -rf ./docs/api && rimraf -rf ./docs/tutorials/*.html && rimraf -rf ./docs/typescript/*.html && rimraf -rf ./docs/*.html && rimraf -rf ./docs/source/_docs && rimraf -rf ./tmp",
+    "docs:clean:legacy": "rimraf index.html && rimraf -rf ./docs/5.x && rimraf -rf ./docs/source/_docs && rimraf -rf ./tmp",
+    "docs:copy:tmp": "mkdirp ./tmp/docs/css && mkdirp ./tmp/docs/js && mkdirp ./tmp/docs/images && mkdirp ./tmp/docs/tutorials && mkdirp ./tmp/docs/typescript && ncp ./docs/css ./tmp/docs/css --filter=.css$ && ncp ./docs/js ./tmp/docs/js --filter=.js$ && ncp ./docs/images ./tmp/docs/images && ncp ./docs/tutorials ./tmp/docs/tutorials && ncp ./docs/typescript ./tmp/docs/typescript && cp index.html ./tmp",
+    "docs:copy:tmp:legacy": "rimraf ./docs/5.x && ncp ./tmp ./docs/5.x",
+    "docs:checkout:gh-pages": "git checkout gh-pages",
+    "docs:checkout:legacy": "git checkout 5.x",
+    "docs:generate": "node ./scripts/website.js",
+    "docs:generate:search": "node docs/search.js",
+    "docs:merge:stable": "git merge master",
+    "docs:merge:legacy": "git merge 5.x",
+    "docs:test": "npm run docs:generate && npm run docs:generate:search",
+    "docs:view": "node ./scripts/static.js",
+    "docs:prepare:publish:stable": "npm run docs:checkout:gh-pages && npm run docs:merge:stable && npm run docs:clean:stable && npm run docs:generate && npm run docs:generate:search",
+    "docs:prepare:publish:legacy": "npm run docs:checkout:legacy && npm run docs:merge:legacy && npm run docs:clean:stable && npm run docs:generate && npm run docs:copy:tmp && docs:checkout:gh-pages && docs:copy:tmp:legacy",
+    "lint": "eslint .",
+    "lint-js": "eslint . --ext .js",
+    "lint-ts": "eslint . --ext .ts",
+    "build-browser": "node ./scripts/build-browser.js",
+    "prepublishOnly": "npm run build-browser",
+    "release": "git pull && git push origin master --tags && npm publish",
+    "release-legacy": "git pull origin 5.x && git push origin 5.x --tags && npm publish --tag legacy",
+    "mongo": "node ./tools/repl.js",
+    "test": "mocha --exit ./test/*.test.js",
+    "test-rs": "START_REPLICA_SET=1 mocha --timeout 30000 --exit ./test/*.test.js",
+    "test-tsd": "node ./test/types/check-types-filename && tsd",
+    "tdd": "mocha ./test/*.test.js --inspect --watch --recursive --watch-files ./**/*.{js,ts}",
+    "test-coverage": "nyc --reporter=html --reporter=text npm test",
+    "ts-benchmark": "ts-benchmark -p ./benchmarks/typescript/simple -f 17/100000 18 29 32",
+    "ts-benchmark-watch": "ts-benchmark -p ./benchmarks/typescript/simple -w ./types -i -s -f 17/100000 18 29 32 -b master"
+  },
   "main": "./index.js",
+  "types": "./types/index.d.ts",
+  "engines": {
+    "node": ">=12.0.0"
+  },
+  "bugs": {
+    "url": "https://github.com/Automattic/mongoose/issues/new"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/Automattic/mongoose.git"
+  },
+  "homepage": "https://mongoosejs.com",
+  "browser": "./dist/browser.umd.js",
   "mocha": {
     "extension": [
       "test.js"
