@@ -66,46 +22,37 @@
     "socks4",
     "socks5"
   ],
-  "license": "MIT",
-  "main": "build/index.js",
-  "name": "socks",
-  "nyc": {
-    "extension": [
-      ".ts",
-      ".tsx"
-    ],
-    "include": [
-      "src/*.ts",
-      "src/**/*.ts"
-    ],
-    "exclude": [
-      "**.*.d.ts",
-      "node_modules",
-      "typings"
-    ],
-    "require": [
-      "ts-node/register"
-    ],
-    "reporter": [
-      "json",
-      "html"
-    ],
-    "all": true
+  "engines": {
+    "node": ">= 10.13.0",
+    "npm": ">= 3.0.0"
   },
-  "private": false,
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/JoshGlazebrook/socks.git"
+  "author": "Josh Glazebrook",
+  "contributors": [
+    "castorw"
+  ],
+  "license": "MIT",
+  "readmeFilename": "README.md",
+  "devDependencies": {
+    "@types/ip": "1.1.0",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^18.0.6",
+    "@typescript-eslint/eslint-plugin": "^5.30.6",
+    "@typescript-eslint/parser": "^5.30.6",
+    "eslint": "^8.20.0",
+    "mocha": "^10.0.0",
+    "prettier": "^2.7.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
+  },
+  "dependencies": {
+    "ip": "^2.0.0",
+    "smart-buffer": "^4.2.0"
   },
   "scripts": {
-    "build": "rm -rf build typings && prettier --write ./src/**/*.ts --config .prettierrc.yaml && tsc -p .",
-    "coverage": "NODE_ENV=test nyc npm test",
-    "coveralls": "NODE_ENV=test nyc npm test && nyc report --reporter=text-lcov | coveralls",
-    "lint": "tslint --project tsconfig.json 'src/**/*.ts'",
     "prepublish": "npm install -g typescript && npm run build",
+    "test": "NODE_ENV=test mocha --recursive --require ts-node/register test/**/*.ts",
     "prettier": "prettier --write ./src/**/*.ts --config .prettierrc.yaml",
-    "test": "NODE_ENV=test mocha --recursive --require ts-node/register test/**/*.ts"
-  },
-  "typings": "typings/index.d.ts",
-  "version": "2.6.2"
+    "lint": "eslint 'src/**/*.ts'",
+    "build": "rm -rf build typings && prettier --write ./src/**/*.ts --config .prettierrc.yaml && tsc -p ."
+  }
 }
