@@ -1,90 +1,49 @@
 {
-  "_from": "statuses@~1.5.0",
-  "_id": "statuses@1.5.0",
-  "_inBundle": false,
-  "_integrity": "sha1-Fhx9rBd2Wf2YEfQ3cfqZOBR4Yow=",
-  "_location": "/statuses",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "statuses@~1.5.0",
-    "name": "statuses",
-    "escapedName": "statuses",
-    "rawSpec": "~1.5.0",
-    "saveSpec": null,
-    "fetchSpec": "~1.5.0"
-  },
-  "_requiredBy": [
-    "/express",
-    "/finalhandler",
-    "/http-errors",
-    "/send"
-  ],
-  "_resolved": "https://registry.npmjs.org/statuses/-/statuses-1.5.0.tgz",
-  "_shasum": "161c7dac177659fd9811f43771fa99381478628c",
-  "_spec": "statuses@~1.5.0",
-  "_where": "/home/lakshy/Desktop/Calling-app/backend/node_modules/express",
-  "bugs": {
-    "url": "https://github.com/jshttp/statuses/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "Jonathan Ong",
-      "email": "me@jongleberry.com",
-      "url": "http://jongleberry.com"
-    }
-  ],
-  "deprecated": false,
+  "name": "statuses",
   "description": "HTTP status utility",
-  "devDependencies": {
-    "csv-parse": "1.2.4",
-    "eslint": "4.19.1",
-    "eslint-config-standard": "11.0.0",
-    "eslint-plugin-import": "2.9.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.7.0",
-    "eslint-plugin-standard": "3.0.1",
-    "istanbul": "0.4.5",
-    "mocha": "1.21.5",
-    "raw-body": "2.3.2",
-    "stream-to-array": "2.3.0"
-  },
-  "engines": {
-    "node": ">= 0.6"
-  },
+  "version": "2.0.1",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
+  ],
+  "repository": "jshttp/statuses",
+  "license": "MIT",
+  "keywords": [
+    "http",
+    "status",
+    "code"
+  ],
   "files": [
     "HISTORY.md",
     "index.js",
     "codes.json",
     "LICENSE"
   ],
-  "homepage": "https://github.com/jshttp/statuses#readme",
-  "keywords": [
-    "http",
-    "status",
-    "code"
-  ],
-  "license": "MIT",
-  "name": "statuses",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/statuses.git"
+  "devDependencies": {
+    "csv-parse": "4.14.2",
+    "eslint": "7.17.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-markdown": "1.0.2",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "8.2.1",
+    "nyc": "15.1.0",
+    "raw-body": "2.4.1",
+    "stream-to-array": "2.3.0"
+  },
+  "engines": {
+    "node": ">= 0.8"
   },
   "scripts": {
     "build": "node scripts/build.js",
     "fetch": "node scripts/fetch-apache.js && node scripts/fetch-iana.js && node scripts/fetch-nginx.js && node scripts/fetch-node.js",
     "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --check-leaks --bail test/",
-    "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
-    "update": "npm run fetch && npm run build"
-  },
-  "version": "1.5.0"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test",
+    "update": "npm run fetch && npm run build",
+    "version": "node scripts/version-history.js && git add HISTORY.md"
+  }
 }
