@@ -1,21 +1,20 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import { CookiesProvider } from "react-cookie";
+import App from "./App";
+import { store } from "./store/index";
+import { Provider } from "react-redux";
 
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-import { store } from './store/index'
-import { Provider } from 'react-redux'
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
+const rootElement = document.getElementById("root");
+const root = createRoot(rootElement);
+
+root.render(
+  <StrictMode>
+    {/* <CookiesProvider> */}
+      <Provider store={store}>
+        <App />
+      </Provider>
+    {/* </CookiesProvider> */}
+  </StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
