@@ -1,41 +1,57 @@
 const jwt = require("jsonwebtoken");
-const refreshModel = require("../Models/refresh-models")
+const refreshModel = require("../Models/refresh-models");
 const accessTokenSecret = process.env.JWT_ACCESS_SECRET;
-const refreshTokenSecret = process.env.JWT_REFRESH_SECRET
+const refreshTokenSecret = process.env.JWT_REFRESH_SECRET;
 
 class TokenService {
-
- generateTokens(payload){
-    const accessToken = jwt.sign(payload, accessTokenSecret ,{
-        expiresIn: "1h"
+  generateTokens(payload) {
+    const accessToken = jwt.sign(payload, accessTokenSecret, {
+      expiresIn:"1m" ,
     });
 
-    const refreshToken = jwt.sign(payload, refreshTokenSecret ,{
-        expiresIn: "1y"
-    })
+    const refreshToken = jwt.sign(payload, refreshTokenSecret, {
+      expiresIn: "1y",
+    });
 
-    return {accessToken ,refreshToken }
+    return { accessToken, refreshToken };
+  }
 
+  async storeRefreshToken(token, userId) {
+    try {
+      await refreshModel.create({
+        token,
+        userId,
+      });
+    } catch (err) {
+      console.log(err.message);
     }
+  }
 
+  async verifyAccessToken(token) {
+    return jwt.verify(token, accessTokenSecret);
+  }
 
-    async  storeRefreshToken(token,userId){
+  async verifyRefreshToken(refreshtoken) {
+    return jwt.verify(refreshtoken, refreshTokenSecret);
+  }
 
-        try{
-           await refreshModel.create({
-                token,
-                userId
-            }) 
-        }catch(err){
-            console.log(err.message)
-        }
+  async findRefreshToken(userId, refreshToken) {
+    return await refreshModel.findOne({
+      userId: userId,
+      token: refreshToken,
+    });
+  }
 
-    }
+  async updateRefreshToken(userId, refreshToken) {
+    return await refreshModel.updateOne(
+      { userId: userId },
+      { token: refreshToken },
+    )
+  }
 
-    async verifyAccessToken(token){
-        return jwt.verify(token , accessTokenSecret )
-    }
+  async removeToken(refreshToken){
+  return await refreshModel.deleteOne({token:refreshToken})
+  }
 }
-
 
 module.exports = new TokenService();
