@@ -1,16 +1,14 @@
-import React, { useState } from 'react'
-import Button from "../../../Components/Shared/Button/Button"
-import Card from "../../../Components/Shared/Card/Card"
-import TextInput from "../../../Components/Shared/TextInput/TextInput"
-import { useDispatch, useSelector } from "react-redux"
-import { setName } from '../../../store/activateSlice'
+import React, { useState } from 'react';
+import Button from '../../../Components/Shared/Button/Button';
+import Card from '../../../Components/Shared/Card/Card';
+import TextInput from '../../../Components/Shared/TextInput/TextInput';
+import { useDispatch, useSelector } from 'react-redux';
+import { setName } from '../../../store/activateSlice';
 
-
-import styles from "./StepName.module.css"
+import styles from './StepName.module.css';
 
 function StepName({ onNext }) {
-
-  const dispatch = useDispatch()
+  const dispatch = useDispatch();
   function nextStep() {
     if (!fullname) {
       return;
@@ -19,16 +17,17 @@ function StepName({ onNext }) {
     dispatch(setName(fullname));
     onNext();
   }
-  const { name } = useSelector((state) => state.activate)
+  const { name } = useSelector((state) => state.activate);
 
   const [fullname, setFullname] = useState(name);
 
   return (
     <>
-    
-      <Card style={styles.card} title="Enter Your Name" icon="name">
-
-        <TextInput value={fullname} onChange={(e) => setFullname(e.target.value)} />
+      <Card style={styles.card} title="Enter Your Name" icon="name" num="5">
+        <TextInput
+          value={fullname}
+          onChange={(e) => setFullname(e.target.value)}
+        />
 
         <div>
           <div className={styles.actionButtonWrap}>
@@ -38,10 +37,9 @@ function StepName({ onNext }) {
         <p className={styles.bottomParagraph}>
           Please enter your name <br /> to let us know about you
         </p>
-
       </Card>
     </>
-  )
+  );
 }
 
-export default StepName
+export default StepName;
